requirements.txt
aiohappyeyeballs==2.4.1
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.2.0
bitarray==2.9.2
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
ckzg==2.0.1
click==8.1.7
colorama==0.4.6
cytoolz==0.12.3
Django==5.1.1
djangorestframework==3.15.2
dnspython==2.6.1
eth-account==0.13.4
eth-hash==0.7.0
eth-keyfile==0.8.1
eth-keys==0.5.1
eth-rlp==2.1.0
eth-tester==0.12.0b1
eth-typing==5.0.0
eth-utils==5.0.0
eth_abi==5.1.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
hexbytes==1.2.1
idna==3.10
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.1.0
narwhals==1.8.3
numpy==2.1.1
packaging==23.2
pandas==2.2.3
parsimonious==0.10.0
pillow==10.4.0
protobuf==5.28.2
psycopg2==2.9.9
py-solc-x==2.0.3
pyarrow==17.0.0
pycryptodome==3.20.0
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
pymongo==4.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyunormalize==16.0.0
pywin32==306
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
rich==13.8.1
rlp==4.0.1
rpds-py==0.20.0
semantic-version==2.10.0
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
streamlit==1.38.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
types-requests==2.32.0.20240914
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
web3==7.3.0
websockets==13.1
yarl==1.12.1
